Design: design LoginForm
