+++ src/style/LoginForm.css
@@ -1,0 +1,65 @@
+.login-container {
+  max-width: 360px;
+  margin: auto;
+  padding: 2rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.login-form {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.login-form input {
+  padding: 0.8rem;
+  font-size: 1rem;
+}
+
+.login-button {
+  padding: 0.8rem;
+  font-size: 1rem;
+  background-color: #333;
+  color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.sns-login {
+  margin-top: 1.5rem;
+  display: flex;
+  justify-content: center;
+}
+
+.naver-button {
+  background-color: white;
+  border: 2px solid #03C75A;
+  border-radius: 50%;
+  width: 60px;
+  height: 60px;
+  padding: 8px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.naver-button img {
+  width: 32px;
+  height: 32px;
+}
+
+.signup-link {
+  margin-top: 1rem;
+  font-size: 0.9rem;
+  text-align: center;
+}
+
+.signup-link a {
+  font-weight: bold;
+  color: #0077cc;
+  text-decoration: none;
+}
